# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SecondHand API",
    "version": "v1",
    "contact": {
      "name": "SecondHand API",
      "url": "https://github.com/sianida26/secondhand-backend/"
    }
  },
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "paths": {
...
    "/bids/accept/{id}": {
      "post": {
        "summary": "Accept Bid",
        "description": "Menerima Tawaran dari calon Buyer",
        "operationId": "handleAcceptBids",
        "security": [
          {
            "BearerTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Bid tidak ditemukan"
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": ["Bids"]
      }
    },
    "/bids/reject/{id}": {
      "post": {
        "summary": "Reject Bid",
        "description": "Menolak Tawaran dari calon Buyer",
        "operationId": "handleRejectBids",
        "security": [
          {
            "BearerTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Bid tidak ditemukan"
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": ["Bids"]
      }
    },
    "/bids/finish/{id}": {
      "post": {
        "summary": "Finish Bid",
        "description": "Menyelesaikan penjualan pada tawaran yang terpilih",
        "operationId": "handleFinishSale",
        "security": [
          {
            "BearerTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Bid tidak ditemukan"
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "tags": ["Bids"]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerTokenAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  }
}